Revert to using mm.instance.asInstanceOf[T]
This method no longer depends on the classpath - any class file can
be inspected and loaded at runtime using the classloader and the
runtime mirrors.  The "clever" bit is telling the classloader to,
uh, load the class.

As a demonstration of this, "DemoResource" is built completely
separately, then gets loaded at run time without the initial
classpath referencing it.  Currently, there's an expectation
that all Sims3ResourceHandler implementations live under "Wrappers"
or "bin" - for testing, I linked "Wrappers" to the bin directory
of the DemoResource project.